Explicit Attachment Size
The Attachment size is an important piece of information for problem reproduction.
In a problem reproduction, one may need to attempt to recreate a similar message/topic/attachment-size. However, the size information is not always available, and the attachment-size may not be just (message-size - header-size), due to binary attachment and meta-data information.